+++ src/buckets/constants/index.ts
@@ -3,3 +3,5 @@ import {startsWith} from 'lodash'
 export const isSystemBucket = (bucketName: string): boolean => {
   return startsWith(bucketName, '_')
 }
+
+export const BUCKET_OVERLAY_WIDTH = 400
